+++ task_01_iryna/src/Algorithms/Algorithm.java
@@ -1,3 +1,5 @@
+package Algorithms;
+
 public abstract class Algorithm {
 
     public abstract void printResult();
