+++ task_01_iryna/src/Main.java
@@ -1,3 +1,5 @@
+import View.SystemInitiator;
+
 public class Main {
 
     public static void main(String[] args) {
@@ -14,5 +16,5 @@ public class Main {
 /*TODO:
     1) сделать валидатор переменных
     2) просто сделать красиво
-    3) третий алгоритм!!
+    3)
 */
