+++ task_01_iryna/src/View/MenuEntry.java
@@ -1,3 +1,5 @@
+package View;
+
 public abstract class MenuEntry {
 
     private String option;
