+++ task_01_iryna/src/View/SystemInitiator.java
@@ -1,8 +1,15 @@
+package View;
+
+import Algorithms.BirthdayAlgorithm;
+import Algorithms.MontyHallAlgorithm;
+import Algorithms.SecretaryAlgorithm;
+import View.Menu;
+import View.MenuEntry;
+
 public class SystemInitiator {
 
 
     public SystemInitiator() {
-
         Menu mainMenu = new Menu();
 
         addFirstMenu(mainMenu);
@@ -12,7 +19,7 @@ public class SystemInitiator {
     }
 
     public void addFirstMenu(Menu menu) {
-        menu.add(new MenuEntry("1","Monty Hall problem") {
+        menu.add(new MenuEntry("1", "Monty Hall problem") {
             @Override
             public void doAction() {
                 MontyHallAlgorithm montyHallAlgorithm = new MontyHallAlgorithm();
@@ -23,7 +30,7 @@ public class SystemInitiator {
     }
 
     public void addSecondMenu(Menu menu) {
-        menu.add(new MenuEntry("2","Birthday problem") {
+        menu.add(new MenuEntry("2", "Birthday problem") {
             @Override
             public void doAction() {
                 BirthdayAlgorithm birthdayAlgorithm = new BirthdayAlgorithm();
@@ -33,7 +40,7 @@ public class SystemInitiator {
     }
 
     public void addThirdMenu(Menu menu) {
-        menu.add(new MenuEntry("3","Secretary problem") {
+        menu.add(new MenuEntry("3", "Secretary problem") {
             @Override
             public void doAction() {
                 SecretaryAlgorithm secretaryAlgorithm = new SecretaryAlgorithm();
